autosuggest component outsourced into own component
