+++ client/src/Components/Pages/Search/index.js
@@ -1,33 +1,28 @@
 import React, { Component } from "react";
 import axios from "axios";
-import Autosuggest from "react-autosuggest";
+
 import { API_SEARCH_URL } from "../../../apiUrls";
-import { ADD_PROFILE_URL } from "../../../constants/naviagationUrls";
+
+import AutosuggestComponent from "./AutoSuggest";
 
 // UI helper functions
 import { SVGCreator, StarRateCreator, SortArrayNewest } from "../../../helpers";
 
 // styles
 import {
-  AutosuggestWrapper,
   HeadlineDiv,
   RowDiv,
   ItemDiv,
   SearchLegendDiv,
   SymbolDiv,
   OrganisationDetailsDiv,
-  AddItemDetails,
   ReviewDetailsDiv,
   InnerDivLastReviews,
-  InnerDivSuggestions,
   ArrowDiv,
   SearchWrapper,
-  SuggestionBox,
-  AddItemBox,
   LegendTitle,
   ReviewsFrame,
   ProfileLink,
-  AddProfileLink,
   ReviewsContainer
 } from "./Search.style";
 
@@ -38,30 +33,10 @@ export const axiosCall = async () => {
   return response;
 };
 
-// gets called when suggestions gets clicked
-export const getSuggestionValue = suggestion => suggestion.name;
-
-export const getSuggestions = (value, organisationsArray) => {
-  const inputValue = value.trim().toLowerCase();
-  const inputLength = inputValue.length;
-  const suggestions = organisationsArray.filter(
-    org => org.name.toLowerCase().slice(0, inputLength) === inputValue
-  );
-
-  // in case there are no suggestions still return a value enabling the 'add' box to be rendered
-  if (suggestions.length === 0) {
-    return [{ isEmpty: true }];
-  }
-
-  return suggestions;
-};
-
 export default class Search extends Component {
   state = {
     isLoading: false,
-    data: null,
-    value: "",
-    suggestions: []
+    data: null
   };
 
   componentDidMount() {
@@ -73,80 +48,6 @@ export default class Search extends Component {
     });
   }
 
-  // functions for autosuggest component
-  // create suggestions array based on user input
-
-  // Autosuggest will call this function every time suggestions are updated
-  onSuggestionsFetchRequested = ({ value }) => {
-    this.setState({
-      suggestions: getSuggestions(value, this.state.data)
-    });
-  };
-
-  // Autosuggest will call this function every time you need to clear suggestions.
-  onSuggestionsClearRequested = () => {
-    this.setState({ suggestions: [] });
-  };
-
-  onChange = (event, { newValue }) => {
-    this.setState({ value: newValue });
-  };
-
-  // render functions
-  // renders individual suggestions in autosuggest search section
-  renderSuggestion = suggestion => {
-    // check if no suggestion available and return so that renderSuggestionsContainer function is still being called (gets deactivated otherwise)
-    if (suggestion.isEmpty) {
-      return null;
-    }
-    return (
-      <ProfileLink to={`/profile/${suggestion._id}`}>
-        <SuggestionBox orgType={suggestion.category}>
-          <InnerDivSuggestions>
-            <SymbolDiv>
-              {SVGCreator("mobile-search-icon")}
-              {SVGCreator(`${organizationIcons[suggestion.category].symbol}`)}
-            </SymbolDiv>
-            <OrganisationDetailsDiv>
-              <h3>{suggestion.name}</h3>
-              <ReviewDetailsDiv>
-                {StarRateCreator(suggestion)}
-                <p>{suggestion.totalReviews} reviews</p>
-              </ReviewDetailsDiv>
-            </OrganisationDetailsDiv>
-            <ArrowDiv>
-              {SVGCreator(`${organizationIcons[suggestion.category].arrow}`)}
-            </ArrowDiv>
-          </InnerDivSuggestions>
-        </SuggestionBox>
-      </ProfileLink>
-    );
-  };
-  // renders all elements and the add item footer
-  renderSuggestionsContainer = ({ containerProps, children, query }) => {
-    if (query && query.length > 0) {
-      return (
-        <div {...containerProps}>
-          {children}
-          <div className="my-suggestions-container-footer" />
-
-          <AddProfileLink
-            to={{ pathname: `${ADD_PROFILE_URL}`, state: { name: `${query}` } }}
-          >
-            <AddItemBox>
-              <InnerDivSuggestions>
-                <SymbolDiv>{SVGCreator("add-item-icon")}</SymbolDiv>
-                <AddItemDetails>
-                  <h3>Add {query}</h3>
-                </AddItemDetails>
-              </InnerDivSuggestions>
-            </AddItemBox>
-          </AddProfileLink>
-        </div>
-      );
-    }
-    return <div {...containerProps}>{children}</div>;
-  };
   // renders last viewed organization section
   renderLastViewed = (org, key) => (
     <ProfileLink key={key} to={`/profile/${org._id}`}>
@@ -171,12 +72,8 @@ export default class Search extends Component {
   );
 
   render() {
-    const { isLoading, value, suggestions, data } = this.state;
-    const inputProps = {
-      placeholder: "🔍 start typing...",
-      value,
-      onChange: this.onChange
-    };
+    const { isLoading, data } = this.state;
+
     if (!isLoading) return <p data-testid="loading">loading...</p>;
 
     return (
@@ -206,18 +103,7 @@ export default class Search extends Component {
             </ItemDiv>
           </RowDiv>
         </SearchLegendDiv>
-        <AutosuggestWrapper>
-          <Autosuggest
-            suggestions={suggestions}
-            onSuggestionsFetchRequested={this.onSuggestionsFetchRequested}
-            onSuggestionsClearRequested={this.onSuggestionsClearRequested}
-            shouldRenderSuggestions={() => true}
-            getSuggestionValue={getSuggestionValue}
-            renderSuggestion={this.renderSuggestion}
-            inputProps={inputProps}
-            renderSuggestionsContainer={this.renderSuggestionsContainer}
-          />
-        </AutosuggestWrapper>
+        <AutosuggestComponent data={data} />
         <HeadlineDiv>
           <p>Or find out what's happening at...</p>
         </HeadlineDiv>
